Improved ArrayRuntimeValue
It now uses a `map[string]IRuntimeValue` to store the data. The keys are transformed to a string (int to str with prefix "i" and use prefix "s" for string).

The array keeps track of the next new key which improves writing speed for entries without key, because the array does not need to iterate all keys anymore.

Fixes #47